+++ const.go
@@ -1,31 +1,14 @@
 package jwt
 
 const (
-	// MethodHS256 is the "alg"
-	// value for HMAC and SHA-256.
-	MethodHS256 = "HS256"
-	// MethodHS384 is the "alg"
-	// value for HMAC and SHA-384.
-	MethodHS384 = "HS384"
-	// MethodHS512 is the "alg"
-	// value for HMAC and SHA-512.
-	MethodHS512 = "HS512"
-	// MethodRS256 is the "alg"
-	// value for RSA and SHA-256.
-	MethodRS256 = "RS256"
-	// MethodRS384 is the "alg"
-	// value for RSA and SHA-384.
-	MethodRS384 = "RS384"
-	// MethodRS512 is the "alg"
-	// value for RSA and SHA-512.
-	MethodRS512 = "RS512"
-	// MethodES256 is the "alg"
-	// value for ECDSA and SHA-256.
-	MethodES256 = "ES256"
-	// MethodES384 is the "alg"
-	// value for ECDSA and SHA-384.
-	MethodES384 = "ES384"
-	// MethodES512 is the "alg"
-	// value for ECDSA and SHA-512.
-	MethodES512 = "ES512"
+	MethodHS256 = "HS256" // Method name for HMAC and SHA-256.
+	MethodHS384 = "HS384" // Method name for HMAC and SHA-384.
+	MethodHS512 = "HS512" // Method name for HMAC and SHA-512.
+	MethodRS256 = "RS256" // Method name for RSA and SHA-256.
+	MethodRS384 = "RS384" // Method name for RSA and SHA-384.
+	MethodRS512 = "RS512" // Method name for RSA and SHA-512.
+	MethodES256 = "ES256" // Method name for ECDSA and SHA-256.
+	MethodES384 = "ES384" // Method name for ECDSA and SHA-384.
+	MethodES512 = "ES512" // Method name for ECDSA and SHA-512.
+	MethodNone  = "none"  // Method name for "none".
 )