+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support for "none" method.
+- Tests for "none" method.
+
 ## [0.3.0] - 2018-02-13
 ### Changed
 - Package structure.
@@ -36,5 +41,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[Unreleased]: https://github.com/gbrlsnchs/jwt/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/gbrlsnchs/jwt/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gbrlsnchs/jwt/compare/v0.1.0...v0.2.0